docs *.mmd to *.md (remove pymmd, using markdown2)

+++ docs/build.py
@@ -1,5 +1,5 @@
 import shutil   
-import pymmd
+import markdown2
 import os,re
 
 template = "resources/template.html"
@@ -7,23 +7,20 @@ files = ["index","step01", "step02", "step03"]
 build = "./build"
 resources = "./resources"
 mdFolder = "./markdown"
-
-images = ".images"
+if os.path.exists(build):
+	shutil.rmtree(build) 
 file = open(template,mode='r')
 templateContent = file.read()
 file.close()
-
-shutil.rmtree(build) 
 destination = shutil.copytree(resources, build)
 for file in files:
 	htmlFile = build + "/" + file + ".html" 
-	mmdFile = mdFolder + "/" + file + ".mmd"
+	mdFile = mdFolder + "/" + file + ".md"
 
 	if os.path.exists(build + "/" + file): os.remove(htmlFile) 
 	#generate html from mmdFile and return body content from HTML
-	print (mmdFile)
-	html_RAW = pymmd.convert_from(mmdFile)
-	body = re.search(r'<body>.*</body>',html_RAW,re.DOTALL).group(0)[6:-7]
+	print (mdFile)
+	body = markdown2.markdown_path(mdFile)
 	body = re.sub(r'<img',r'<br /> <img class="img-fluid"',body)
 	if file=="index":
 		templateIndex = re.sub(r'<a href="index.*</a>',r'',templateContent,re.DOTALL) if file=="index" else body
